project_root/requirements.txt
numpy
Pillow
scikit-learn
scikit-image
matplotlib